Shifted home page API call to ProductConatiner
